adding get and verify endpoints
